Revert "Rewrite text squashing and text wrapping"
This reverts commit 1a9ba27397c92a4301c0c80e451a39be4aa7a650.
