--- src/get-max-width.js
@@ -1,3 +1,0 @@
-export default yogaNode => {
-	return yogaNode.getComputedWidth() - (yogaNode.getComputedPadding() * 2);
-};
